Split out hoverfly wrapper into new files
