+++ hoverctl/wrapper/hoverfly.go
@@ -1,7 +1,6 @@
 package wrapper
 
 import (
-	"bytes"
 	"crypto/tls"
 	"encoding/json"
 	"errors"
@@ -65,222 +64,6 @@ type MiddlewareSchema struct {
 
 type ErrorSchema struct {
 	ErrorMessage string `json:"error"`
-}
-
-// Wipe will call the records endpoint in Hoverfly with a DELETE request, triggering Hoverfly to wipe the database
-func DeleteSimulations(target configuration.Target) error {
-	response, err := doRequest(target, "DELETE", v2ApiSimulation, "", nil)
-	if err != nil {
-		return err
-	}
-
-	defer response.Body.Close()
-
-	if response.StatusCode != 200 {
-		return errors.New("Simulations were not deleted from Hoverfly")
-	}
-
-	return nil
-}
-
-// GetMode will go the state endpoint in Hoverfly, parse the JSON response and return the mode of Hoverfly
-func GetMode(target configuration.Target) (string, error) {
-	response, err := doRequest(target, "GET", v2ApiMode, "", nil)
-	if err != nil {
-		return "", err
-	}
-
-	defer response.Body.Close()
-
-	apiResponse := createAPIStateResponse(response)
-
-	return apiResponse.Mode, nil
-}
-
-// Set will go the state endpoint in Hoverfly, sending JSON that will set the mode of Hoverfly
-func SetModeWithArguments(target configuration.Target, modeView v2.ModeView) (string, error) {
-	if modeView.Mode != "simulate" && modeView.Mode != "capture" &&
-		modeView.Mode != "modify" && modeView.Mode != "synthesize" {
-		return "", errors.New(modeView.Mode + " is not a valid mode")
-	}
-	bytes, err := json.Marshal(modeView)
-	if err != nil {
-		return "", err
-	}
-
-	response, err := doRequest(target, "PUT", v2ApiMode, string(bytes), nil)
-	if err != nil {
-		return "", err
-	}
-
-	if response.StatusCode == http.StatusBadRequest {
-		return "", handlerError(response)
-	}
-
-	apiResponse := createAPIStateResponse(response)
-
-	return apiResponse.Mode, nil
-}
-
-// GetDestination will go the destination endpoint in Hoverfly, parse the JSON response and return the destination of Hoverfly
-func GetDestination(target configuration.Target) (string, error) {
-	response, err := doRequest(target, "GET", v2ApiDestination, "", nil)
-	if err != nil {
-		return "", err
-	}
-
-	defer response.Body.Close()
-
-	apiResponse := createAPIStateResponse(response)
-
-	return apiResponse.Destination, nil
-}
-
-// SetDestination will go the destination endpoint in Hoverfly, sending JSON that will set the destination of Hoverfly
-func SetDestination(target configuration.Target, destination string) (string, error) {
-	response, err := doRequest(target, "PUT", v2ApiDestination, `{"destination":"`+destination+`"}`, nil)
-	if err != nil {
-		return "", err
-	}
-
-	apiResponse := createAPIStateResponse(response)
-
-	return apiResponse.Destination, nil
-}
-
-// GetMiddle will go the middleware endpoint in Hoverfly, parse the JSON response and return the middleware of Hoverfly
-func GetMiddleware(target configuration.Target) (v2.MiddlewareView, error) {
-	response, err := doRequest(target, "GET", v2ApiMiddleware, "", nil)
-	if err != nil {
-		return v2.MiddlewareView{}, err
-	}
-
-	defer response.Body.Close()
-
-	middlewareResponse := createMiddlewareSchema(response)
-
-	return middlewareResponse, nil
-}
-
-func SetMiddleware(target configuration.Target, binary, script, remote string) (v2.MiddlewareView, error) {
-	middlewareRequest := &v2.MiddlewareView{
-		Binary: binary,
-		Script: script,
-		Remote: remote,
-	}
-
-	marshalledMiddleware, err := json.Marshal(middlewareRequest)
-	if err != nil {
-		return v2.MiddlewareView{}, err
-	}
-
-	response, err := doRequest(target, "PUT", v2ApiMiddleware, string(marshalledMiddleware), nil)
-	if err != nil {
-		return v2.MiddlewareView{}, err
-	}
-
-	err = handleResponseError(response, "Hoverfly could not execute this middleware")
-	if err != nil {
-		return v2.MiddlewareView{}, err
-	}
-
-	apiResponse := createMiddlewareSchema(response)
-
-	return apiResponse, nil
-}
-
-func GetLogs(target configuration.Target, format string) ([]string, error) {
-	headers := map[string]string{
-		"Content-Type": "text/plain",
-	}
-
-	if format == "json" {
-		headers["Content-Type"] = "application/json"
-	}
-
-	response, err := doRequest(target, "GET", v2ApiLogs, "", headers)
-	if err != nil {
-		return []string{}, err
-	}
-
-	defer response.Body.Close()
-
-	responseBody, _ := ioutil.ReadAll(response.Body)
-	if format == "json" {
-		var logsView v2.LogsView
-
-		err := json.Unmarshal(responseBody, &logsView)
-		if err != nil {
-			return nil, err
-		}
-
-		var logs []string
-		for _, log := range logsView.Logs {
-			jsonLog, err := json.Marshal(log)
-			if err != nil {
-				return nil, err
-			}
-
-			logs = append(logs, string(jsonLog))
-		}
-
-		return logs, nil
-	} else {
-		return strings.Split(string(responseBody), "\n"), nil
-	}
-}
-
-func ImportSimulation(target configuration.Target, simulationData string) error {
-	response, err := doRequest(target, "PUT", v2ApiSimulation, simulationData, nil)
-	if err != nil {
-		return err
-	}
-
-	if response.StatusCode != 200 {
-		body, _ := ioutil.ReadAll(response.Body)
-		var errorView ErrorSchema
-		json.Unmarshal(body, &errorView)
-		return errors.New("Import to Hoverfly failed: " + errorView.ErrorMessage)
-	}
-
-	return nil
-}
-
-func FlushCache(target configuration.Target) error {
-	response, err := doRequest(target, "DELETE", v2ApiCache, "", nil)
-	if err != nil {
-		return err
-	}
-
-	if response.StatusCode != 200 {
-		return errors.New("Cache was not set on Hoverfly")
-	}
-
-	return nil
-}
-
-func ExportSimulation(target configuration.Target) ([]byte, error) {
-	response, err := doRequest(target, "GET", v2ApiSimulation, "", nil)
-	if err != nil {
-		return nil, err
-	}
-
-	defer response.Body.Close()
-
-	body, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		log.Debug(err.Error())
-		return nil, errors.New("Could not export from Hoverfly")
-	}
-
-	var jsonBytes bytes.Buffer
-	err = json.Indent(&jsonBytes, body, "", "\t")
-	if err != nil {
-		log.Debug(err.Error())
-		return nil, errors.New("Could not export from Hoverfly")
-	}
-
-	return jsonBytes.Bytes(), nil
 }
 
 func createAPIStateResponse(response *http.Response) APIStateSchema {
